refactor(cp): move part of command options for cpenv to core
